// NextMigration/lib/api/types.ts
export interface CreateClassroom {
    name: string;
    type: string;
}

export interface UpdateClassroom {
    name?: string;
    owner?: string;
    type?: string;
    verified?: boolean;
}

export interface Classroom {
    id: string;
    name: string;
    owner_id: string;
    type: string;
    created_at: string;
    verified: boolean;
}

export interface ClassroomSet {
    set_id: string;
    set_type: string;
    classroom_id: string;
    added_by: string;
}

export interface FullClassroomSet extends ClassroomSet {
    title: string;
    course: string;
    owner: string;
    created_at: string;
}

export interface ClassroomUser {
    user_id: string;
    classroom_id: string;
    img_url: string;
    role: string;
    joined_at: string;
    displayName: string;
    streak: number;
    verified: boolean;
}

export interface FullClassroom extends Classroom {
    sets: FullClassroomSet[];
    users: ClassroomUser[];
}

export interface ClassroomListResponse {
    classrooms: Classroom[];
}

export interface CreateClassroomUser {
    classroom_id: string;
    role: string;
}

export interface UpdateClassroomUser {
    role: string;
}

export interface CreateClassroomSet {
    set_id: string;
}

export interface CreateClassroomSets {
    sets: string[];
}